defconfig: enable CONFIG_MEMCG and unset CONFIG_MEMCG_DISABLED
Even after enabling memory controller in defconfig, it was found out that it gets disabled automatically on system.img which was found out by lxc check config. A similar issue was found in debian which was reported as bug. To fix this issue we are supposed to add CONFIG_MEMCG which helps the memory controller to remain enabled. But it was also found that CONFIG_MEMCG gets automaticaly disabled too.
So, CONFIG_MEMCG_DISABLED is unset so that it can't disable CONFIG_MEMCG. We also had a option to add "cgroup_enable=memory" without quotes in BoardConfigCommon.mk in kernel cmdline but as adding and unsetting did job for us, we skipped adding that entry.